angle solver qsp/qsvt (#6483)
**Context:**

This PR appears in the context of the QSVT project, specifically
focusing on the implementation of the angle solver, functionality to
calculate the angles needed to apply a polynomial transformation.

**Description of the Change:**

Added `qml.poly_to_angles`, the main functionality of the PR that takes
as input a polynomial and returns the angles.
Also included is the function `qml.transform_angles`, which transforms
the angle format between qsp and qsvt.
Both these functions and other auxiliary functions have been added in
`qsvt.py`.

**Benefits:**

Calculating angles is not a simple task and existing libraries have some
limitations. With this new functionality we will facilitate the use of
routines such as qsp and qsvt.

**Possible Drawbacks:**

Currently only one solver has been introduced. In the future it is
expected to add more that will work for more extreme cases: larger
polynomials (degree > 1000) or degenerate polynomials


[sc-72631]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>