Update documentation of `TrotterProduct` regarding `operands` and circuit structure (#6629)
**Context:**
The decomposition logic of `TrotterProduct` relies on the `operands`
attribute of the input `hamiltonian` argument.
The documentation makes it sound like the `terms()` method of the
`hamiltonian` is the used information to determine the circuit structure
of the decomposition.
Also, there is no example illustrating the logic behind these
terms/operands being grouped when creating the circuit from
`TrotterProduct`.

**Description of the Change:**
- Change wording from `terms` towards `operands`
- Add example illustrating how the `operands` impact the created
circuit.

**Benefits:**
Documentation clarity/extent

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**

[sc-79003]

---------

Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>