Transforms can be captured with PLxPR (#6633)
[sc-72804] [sc-72803]

This PR adds a primitive for capturing transforms.

**Description of change**
* Add a primitive to capture transforms: every time a transform is
created, it automatically gets registered as a primitive
* This primitive is currently only for capturing transforms and does not
actually transform the input function.
* This primitive gets bound (binded?) when calling the transform on the
QNode or a function.
* When applied to a function, the transform is captured as a
higher-order primitive. Evaluating the subsequent JAXPR doesn't do
anything (i.e, the inner JAXPR is evaluated without any transformations)
* When applied to a QNode, the transform is appended to the QNode's
`TransformProgram` as well so that executing QNodes applies the
transform(s) to the constructed tape, and also captured as a primitive.
* Add `plxpr_transform` attribute to `TransformDispatcher` and
`TransformContainer`. This is not currently used and is being added here
to set up scaffolding in `pennylane.transfoms.core` for #6389 .

**Benefits**
Transforms can be captured into PLxPR and no extra work is required when
creating new transforms

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>