Add autograph support for `for` loops (#6426)
**Context:**
We added initial autograph support with custom PennyLane conversion for
if/else statements in #6406, and `while` loops in #6413. This PR adds
support for `for` loops as well.

**Description of the Change:**
We add a custom PennyLane implementation for for_stmt for autograph,
based on the existing Catalyst implementation.

**Benefits:**

We can have logic like this in a captured function:

```
@qml.qnode(qml.device("default.qubit", wires=1))
def f(params):
    for x in params:
        qml.RY(x, wires=0)
    return qml.expval(qml.PauliZ(0))

ag_circuit = run_autograph(f)
jaxpr = jax.make_jaxpr(ag_circuit)(jnp.array([1.0, 2.0, 3.0]))
```
and the resulting JAXPR will contain a for-loop instead of unrolling the
entire loop.

**Possible Drawbacks:**
Not a drawback, per se, but this is the part of the implementation that
is most "behind" catalyst in terms of feature parity. Catalyst has an
option to fall back on a Python implementation that is based on queueing
instead of raising an error in some circumstances. A few test are
`xfailed` with a note that they are implemented in Catalyst.

**Relevant Shortcut Stories:**
[sc-71821]

---------

Co-authored-by: andrijapau <andrija.paurevic@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>