Add simulate of analytic case for DefaultMixedNewAPI (#6618)
**Context:**
The implementation of method `simulate` is the most important one under
the architecture of new API standard. The point of `simulate`, is
basically abstract both the analytic version (controlled by
`shots=None`, or infinite shots in a fancier way) and the MC version
(finite shots). Meanwhile, the functionality of such method should not
truly depend on our details of qml implementationi; they better only
rely on either the fundamental computation libs, e.g. math lib and qnp,
or the most essential infras, e.g. Tape/Scripts and Operations.

**Description of the Change:**
 - Added `qml.devices.qubit_mixed.simulate`
 - Added the analytic branch of `simulate`
 - Added the corresponding test suites

**Benefits:**
Implementation of this submodule is necessary to several other PR's

**Possible Drawbacks:**
finite-shot branch not yet implemneted

**Related GitHub Issues:**
[sc-73320]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>