[Capture] remove x64 mode setting in tests (#6671)
**Context:**

The PR making so that the jaxpr is natively executed on the device
(#6596) was getting test failures due to a global state change in a
different part of the test suite. This was tracked down to be due to the
jax config getting updated so much when running the test suite. This PR
gets rid of all the setting and re-setting of the jax config, leading to
the removal of these confusing failures.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**