[Capture] Qnode natively executes jaxpr on the device (#6596)
**Context:**

Though we can now capture entire workflows and execute jaxpr on default
qubit, we are still falling back to the old execution pipeline for
execution.

**Description of the Change:**

Adds a new implementation to the qnode primitive that sends the qfunc
jaxpr to the device.


It also adds a couple minor fixes to `DefaultQubitInterpreter`.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-78422]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>