`Controlled` subclasses no longer bind base class primitives on initialization (#6672)
The constructors of subclasses of `Controlled` (`CNOT`, `CRX`, etc.)
construct the base operator so that the parent `Controlled` class can be
initialized correctly. However, in doing so, the primitive of the base
operator is also binded. We don't want this, because calling
constructors is part of the concrete implementation, and binding
primitives can create abstract values (when using `make_jaxpr`, etc.),
which we don't want.

Changes:
* Update all currently existing subclasses of `Controlled` to use
`type.__call__(base_op_class, ...)` instead of `base_op_class(...)` to
construct the base operator. This skips the call to `op._primitive.bind`
and just directly creates the class we want to construct.

Downsides:
* This will need to be done for all subclasses of `Controlled`, so it is
an additional detail that devs have to keep in mind when adding new
controlled ops.

[sc-79769]