Standardize execution interfaces by adding internal `Enum` object (#6643)
**Context:**  

This is a follow-up to #6649.  

The purpose of this PR is to standardize how the `interface` value is
handled, preventing it from being remapped multiple times during
execution. Users should provide their preferred interface, which will be
mapped directly to a canonical value for consistent use. If the
interface is set to "auto", it will be inferred from the data type and
converted into a supported canonical interface. This inferred interface
value will then be used to configure the execution later down the
pipeline.

**Description of the Change:**  

- Updated `qnode.py` and `execution.py` to use a new canonical interface
`Enum`.
- Modified tests that previously expected `interface` to be a string in
parts of the code where it is now represented as an enum value.
- Removed support for `interface=None`. Instead, the default interface
is now explicitly set to `numpy`.
- Adjusted `execution config` to store only the canonical interface
value.

**Benefits:**  

- Enhances code testability.  
- Resolves ambiguities in how the `interface` value is handled during
execution.

**Possible Drawbacks:**  None identified.  

[sc-73432]  
[sc-79735]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>