[BugFix] Removing the Hermitian check in `qml.Hermitian` (#6642)
**Context:** After some discussion, we decided to completely remove the
'Hermitianity' input validation check in the `qml.Hermitian` class since
it might be expensive for very large matrices. Furthermore, it is used
every time we call the `compute_matrix` and `compute_decomposition`
methods.

Our current validation strategy is to raise errors if the corresponding
error is rather difficult to debug and/ or the sanity check is
sufficiently cheap and jittable.

In this case, the validation check is neither cheap nor jittable. And
the user might actually want to provide a non-hermitian matrix-based
observable.

**Description of the Change:** As above.

**Benefits:** Faster execution and jittability.

**Possible Drawbacks:** The main drawback is that if the user provides
by accident a non-Hermitian matrix to the `qml.Hermitian` class
believing it is actually Hermitian, he/she might experience strange
errors because the code doesn't check explicitly the input matrix
anymore.

**Related GitHub Issues:** #6640 

**Related Shortcut Stories:** [sc-79122]