Move `tests/interfaces/*` to new folders in `tests/workflow/interfaces` (#6704)
**Context:**

As part of the `workflow` clean-up epic, we are doing some organization
and improvement of tests.

**Description of the Change:**

Purely a file relocation PR:
1. Move `tests/interfaces/*` to `tests/workflow/interfaces`.
2. Split relevant files into three folders: `run`, `execute` and
`qnode`.

Unintended side-effect: 

When the files were moved, a new seed was generated for the tests,
causing some failures. To address this, the tests were checked for
correct behavior and salted. However, salting didn’t resolve one issue,
so the approach was adjusted by switching from `rtol` to `atol`. This
change was necessary because specifying `rtol` for values near zero is
ineffective when the noise is absolute rather than proportional to the
value.

**Benefits:** Improve organization of `workflow` tests.

**Possible Drawbacks:** None

[sc-80318]