Create a `map_wires` transform that transforms PLxPR (#6697)
**Context:** This PR implements a `MapWires` transform that can be used
to map the wires of a `QNode` with capture enabled.

The architecture for capturing transforms recently changed from
`Tracers` to a `PlxprInterpreter`-style implementation for each
transform. This has been brought up in several discussions involving
Product.

For transparency, I list here all this story's relevant features to
avoid future ambiguities.

- `map_wires` is a function, not a transform, which internally uses the
`_map_wires_transform` quantum transform.

- The `MapWiresInterpreter` implemented in this PR only receives the
`wire_map` argument, while the `_map_wires_transform` transform can also
receive a `queue` argument, to determine whether or not to queue the
object when recording. The usefulness of this parameter in the capture
context is questionable, so it has not been taken into account due to
lack of time.

- For the UI, as discussed with Product, this PR follows the same UI
used for `decompose` and `cancel_inverses`. How to define a better UI
for transforms when capture is enabled will be one of the goals for Q1
2025.

**Description of the Change:** As above.

**Benefits:** We can map the wires of `QNode`s with capture enabled.

**Possible Drawbacks:** None that I can think of. Limitations due to the
current approach (based on `PlxprInterpreter` rather than `Tracers`, as
originally planned) have been discussed extensively internally and with
the Product team.

**Related GitHub Issues:** None.

**Related Shortcut Stories:** [sc-72845]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>