Wires object can be converted to JAX array (#6699)
**Context:**
Iterating over a `Wires` object fails with AutoGraph conversion, because
it's not a valid AutoGraph `iteration_target`. This means we need to do
things like `for w in dev.wires.labels` instead of just `for w in
dev.wires`.

**Description of the Change:**
We add a `__jax_array__` method on the `Wires` class. 

**Benefits:**
A `Wires` object can now be converted using `jax.numpy.asarray(wires)`,
and works with AutoGraph, so we can iterate over it for program capture.
This *should* also fix the same issue in Catalyst with AutoGraph.

[sc-80256]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>