Create `expand_plxpr_transforms` function for unwrapping transforms natively in plxpr (#6722)
[sc-80562]

**Note to reviewers:** Please don't be put off by the number of lines 😅
. Half of the new lines are just moving the `DecomposeInterpreter`,
`MapWiresInterpreter`, and `CancelInversesInterpreter` from the
`capture.transforms` to the `transforms` module.

**Context:**
This PR adds a function `qml.capture.expand_plxpr_transforms` that can
be used to actually apply transform primitives to plxpr. An example is
shown in the comment below of how it "unravels" transforms. This PR also
reorganized the transforms program capture. See before for details.

**Description of the Change:**
* Add a `qml.capture.expand_plxpr_transforms` function which returns a
new function that applies all transforms that are present as primitives
in the original function.
* Added an `ExpandTransformsInterpreter`. This interpreter does not do
anything special, but all PL transforms' primitives get automatically
registered so that the interpreter has a custom handler for evaluating
transform primitives. This custom implementation "unravels" transforms
as they are getting evaluated.
* Update the `plxpr_transform` argument of `qml.transform`. This
argument should now be a function with the following signature:
  ```python
def dummy_plxpr_transform(jaxpr: jax.core.Jaxpr, consts: list, targs:
list, tkwargs: dict, *args) -> jax.core.ClosedJaxpr:
      ...
  ```
By doing so, we can essentially generalize how plxpr transforms can be
implemented. We chose to use interpreters because of their ease of use,
but as long as the `plxpr_transform` function follows the "jaxpr to
jaxpr" signature, the actual implementation of the transform can be
whatever we want it to be 😄
* Reorganized code:
  * Remove `capture.transforms` submodule.
* Move all interpreters to the same files as their respective
transforms.
* These interpreters are placed inside a function so that we can work
around whether or not jax is installed. This is consistent with how we
handle the creation of certain classes/functions/primitives in the
`qml.capture` submodule.

**Benefits:**
Plxpr transforms are integrated with the existing transforms API

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>