Clean-up `Wires` object usage across source code (#6689)
**Context:**

Follow up to #6713.

It is not possible to initialize a `Wires` object with `wires=None`
after #6713. However, we should not be using `wires=None` in our
code-base unless it's appropriate.

**Description of the Change:**

This PR reviews the code-base to ensure that `wires=None` does not occur
where it is not applicable. When appropriate, it updates the default to
`wires=()` or uses no default at all.

**Benefits:** Code clean-up.

**Possible Drawbacks:** None identified.

[sc-78998]

---------

Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>