add qml.math.grad and qml.math.jacobian for differentiating any interface (#6741)
**Context:**

We often have to rewrite tests four different ways to test
differentiability for each different interface. This is rather
time-intensive and annoying. By adding and using a `qml.math.grad`, we
can write one test that applies to all interfaces.

**Description of the Change:**

Adds a `qml.math.grad` and `qml.math.jacobian` with a jax-like interface
that can differentiate with respect to any argument inputs.

**Benefits:**

**Possible Drawbacks:**

We have several different options for interface, but I decided to do my
best to mirror the jax interface.

**Related GitHub Issues:**

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>