Bugfix for wire_options (#6805)
**Context:**
When using wire labels that aren't integers ordered from largest to
smallest (like strings, or a wire_order that isn't "in order"), the
custom wire options are applied to the incorrect wire.

```
dev = qml.device('default.qubit', wires=4)

@qml.qnode(dev)
def circuit():
    for w in dev.wires:
        qml.X(w)
    return qml.expval(qml.Z(0))

wire_options = {'color': 'teal',
                'linewidth': 5,
                3: {'color': 'orange', 'linestyle': '--'}, # wire 3 should be orange and dashed
}
fig, ax = qml.draw_mpl(circuit, wire_order=[1, 3, 0, 2], wire_options=wire_options)()
```

This should make wire 3 orange, but instead makes the wire at _index_ 3
orange:

![image](https://github.com/user-attachments/assets/193925cb-ee97-43e7-9d3a-22b0da700589)

**Description of the Change:**
We update it to fix the bug. This involves modifying the signature of
`MPLDrawer`.

**Benefits:**
It works as expected now.

**Possible Drawbacks:**
While MPLDrawer is internal, and even other PL libraries should be using
`draw_mpl` (whose signature is unchanged), this is technically a small
breaking change. However, a search in Catalyst, Lightning and QML
confirms that none of our other packages are calling this, and the only
tests that needed modification were the ones directly testing the class,
so it seems safe to say we've kept it very isolated and this shouldn't
cause any problems.