Allow device to configure conversion to numpy and use of `pure_callback` (#6788)
**Context:**

While we have logic for sampling with jax, it does not really integrate
very well into the workflow. While you can technically set
`diff_method=None` right now and jit the execution end-to-end, trying to
jit `diff_method=None` will cause incomprehensible error messages on
non-DQ devices.

We want to *forbid* differentiation `diff_method=None`, but keep a way
to jit a finite shot execution.

**Description of the Change:**

In order to allow jitting finite shot executions, we need a way for the
device to be able to configure whether or not the data is converted to
numpy. To do so, we simply add another property to the
`ExecutionConfig`, `convert_to_numpy`. If `False`, then we will not use
a `pure_callback` to convert the parameters to numpy. If `True`, we use
a `pure_callback` and convert the parameters to numpy.

**Benefits:**

Speed ups due to being able to jit the entire execution.


![image](https://github.com/user-attachments/assets/738076c6-7bb5-4c38-a8cc-97e138325dbc)


**Possible Drawbacks:**

`ExecutionConfig` gets an addtional property, making it more
complicated.

**Related GitHub Issues:**

Fixes #6054 Fixes #3259  Blocks #6770

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>