Remove string support for `control_values` of `MutliControlledX` (#6835)
**Context:**
In PL0.36 we deprecated the string type support for argument
`control_values` of the operator `MultiControlledX` initialization, e.g.
`control_values='01'` shall not be allowed anymore. In this PR we would
like to remove this entirely with a strict type check.

**Description of the Change:**
- Added a type check with `isinstance` to raise `ValueError` if the
`control_values` does not follow the designed pattern.
- Added a type hint at correpsonding position within the init func
signature
 - Removed a test that presumes the validity of str input
 - Alter the deprecation check into error check

**Benefits:**
Less redundancy

**Possible Drawbacks:**
Potential existence of uncovered indirect usage of string input in
`MultiControlledX`

**Related GitHub Issues:**
[sc-81527]