Update docs workflow schedule (#6867)
[sc-82706]

Docs workflow currently opens a PR to update stable deps on Wednesday,
while the tests workflow opens the stable deps update PR on Monday.
Having 2 PRs per week for the same thing is annoying, so the schedule
time for the docs workflow should be the same as the tests workflow.