Bump `torch` version in CI to 2.5.0 (#6868)
As name says. We should be testing against the latest version of
`torch`. This PR updates the torch version used in CI to `2.5.0`.

Also updated torch installation to use `~=` instead of `==` for choosing
the version, so that bug fix releases (highest possible `2.5.X`) are
automatically used instead of sticking with `2.5.0`. This should be a
safe change to make, as torch only adds bug fixes to patch releases and
no breaking changes
([ref](https://lightning.ai/docs/pytorch/stable/versioning.html)).