[Capture] Add backprop validation (#6852)
**Context:**

We currently use un-validated backprop for differentiation with program
capture. This leads to some unintuitive errors if you try and take a
gradient on lightning with capture enabled.

**Description of the Change:**

Adds some validation to make sure the device supports backprop. Adds the
backprop logic to a `_backprop` jvp function, and dispatches to that
method based on the diff method.

**Benefits:**

Improved error messages when backprop or the requested diff method isn't
supported.

**Possible Drawbacks:**

The code currently is a little clunky, but it is private so we should be
able to move things around once we have more information.

**Related GitHub Issues:**

[sc-82166]