Revert end-to-end jitting with default qubit (#6869)
**Context:**

In #6788 , we started allowing executions on default qubit to be jitted
from end-to-end. Unfortunately, we found that the compilation overheads
on these executions can get very, very expensive. So until we find a way
to reduce the compilation overheads, we are using pure callbacks and
conversion to numpy.

**Description of the Change:**

Default to `convert_to_numpy=False`, and xfail relevant tests. This
change can be undone once we figure out how to resolve the compilation
issue.

**Benefits:**

Reduced compilation overheads, because the execution itself does not get
compiled.

**Possible Drawbacks:**

Slow down on post-compiled workflows. No way to jit an entire execution
on default qubit.

**Related GitHub Issues:**

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>