Add merge queue trigger for required workflows (#6860)
**Context:**
This pull request adds a new trigger to the existing workflows that run
`on.pull_request`. This trigger indicates to GitHub which workflows
needs to be run when a merge queue is building.

**Description of the Change:**
The change is adding `on.merge_group` to our required workflows.

**Benefits:**
This change itself will not enable merge queues, that needs to be
enabled from the admin settings of branch protection rules.

The changes in this PR mainly tell merge queues which workflows to run.

**Possible Drawbacks:**
Usage of merge queue is a new thing for pennylane, if issues arise we
can rollback.

**Related GitHub Issues:**
None. [sc-82039]