fix interface detection with closure variables (#6892)
**Context:**

We were not properly detecting the interface of a qnode when all
parameters were being passed as closure variables:
```
import jax
import numpy as np

V = jax.numpy.array([[ 0.53672126+0.j        , -0.1126064 -2.41479668j],
              [-0.1126064 +2.41479668j,  1.48694623+0.j        ]])
eigen_vals, eigen_vecs = jax.numpy.linalg.eigh(V)
umat = eigen_vecs.T
wires = range(len(umat))

@jax.jit
@qml.qnode(qml.device("lightning.qubit", wires = wires))
def circuit():
   qml.BasisRotation(wires=wires, unitary_matrix=umat)
   return qml.state()

print(circuit())
```

Since with the qnode, we detect the interface from what it gets called
with, we did not detect that the `umat` was a jax variable, and that
it's decomposition would produce tracers.

**Description of the Change:**

We stop detecting the interface from what the qnode is called with. We
only detect the interface from what is on the tape.

We also detecting jitting by creating a new array and seeing if it a
tracer. Closure variables are not tracers until they are transformed, so
detecting jitting with unmodified closure variables will not work.

**Benefits:**

Better use of closure variables.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>