Deprecate `QubitUnitary` as a valid input to `ControlledQubitUnitary` (#6840)
**Context:**
Providing a `QubitUnitary` as the base operator defeats the purpose
entirely and is just a messier way of using
`qml.ctrl(qml.QubitUnitary(...), ...)`

For richer context, check
https://github.com/PennyLaneAI/pennylane/pull/6839 for better
illustration; this PR should in principle be an add-on to it.

Note: this deprecation is related with a previous bug. Trace via
attached story link to see full discussions.

**Description of the Change:**
`__init__` and `_unflatten`: If get a `QubitUnitary` base, raise PLDW
and then get matrix by `base.matrix()`.

Added an overloading method `_flatten`. This is a fix to
https://github.com/PennyLaneAI/pennylane/pull/6839.

Deleted two tests that restrictively replied on the support of
`QubitUnitary` base.



**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-80843]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>