Improve approximate set generation in `qml.ops.sk_decomposition` (#6855)
**Context:** The approximate set for the Solovay-Kitaev decomposition is
currently prepared by populating a trie-based data structure (TbDS) with
a BFS traversal. This is efficient but misses out on a few future gate
sequences because some parent sequences are caught in the equality check
condition.

**Description of the Change:**
1. To allow traversal over some potential missed-out candidate
sequences, we relax the equality check to include information regarding
the tracked global phases. This unlocks.
2. While the previous step unlocks the previously forbidden parent
sequences, it also adds redundant sequences in the set, which is
undesirable. So we prune our TbDS by finding the redundant pool
sequences and ranking them based on their T-gate count.
3. To perform the T-gate ranking efficiently, we also build and store
this information for each sequence during the traversal itself.
4. To unlock higher `max_length` for broader TbDS, we now separate
checking for equivalence with a local and global `KDTree`. The local one
is updated as the TbDS is being constructed for a given depth, whereas
the global one is updated after the construction for a given depth.

**Benefits:** This makes the approximate set decently dense than before,
unlocking efficient decomposition for some previously missing edge
cases.

**Possible Drawbacks:** This adds a constant overhead incurred from
traversing additional parent nodes and post-build pruning. However, this
is negligible; as the approximate set once built is cached.

**Related GitHub Issues:** [sc-82463] #6131