Fix return type call in noise conditionals (#6910)
**Context:**
Even though the recent fix
https://github.com/PennyLaneAI/pennylane/commit/5910ef7b6d5bb16676fb5a3a7f75a264200ebbca
successfully removed the warnings in test suite of master checking, the
CI from the PR that brings back the warning-as-errors
[indicates](https://github.com/PennyLaneAI/pennylane/actions/runs/13074879268/job/36484511004?pr=6880)
that there are still deprecation warnnigs raised.

**Description of the Change:**
Completely remove the use of any `return_type` usages in `noise` module.

**Benefits:**
Clean up deprecation remains

**Possible Drawbacks:**
~Hard to test right now. CI doesn't work due to lack of
warning-as-error. One has to test locally to see the impact.~
Tested via merging the branch that was actively fixing CI no
warning-as-error issue

**Related GitHub Issues:**

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: andrijapau <andrija.paurevic@xanadu.ai>