Update `interface-unit-tests.yml` handling of `pytest` warnings (#6880)
CI is hiding deprecation warnings, because the use of `-W` (introduced
to support the `warnings-as-errors` workflow) overrides `pytest.ini`.
This PR moves the creation of the pytest arg to a separate job, and uses
the output of this job as the pytest arg. This way, we can avoid using
`-W` altogether if the warnings-as-errors workflow isn't running.

[sc-83043]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: andrijapau <andrija.paurevic@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>