Add FTQC bare module (#6912)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** Adds an experimental `ftqc` module to PennyLane.

**Description of the Change:** Added new bare module `ftqc`. Added
updates to `tach.toml` to ensure use of `ftqc` module does not propagate
outside the module, or a user example:
e.g. adding `from pennylane.ftqc import *` to `operation.py` yields:
```bash
❌ pennylane/operation.py[L242]: Cannot import 'pennylane.ftqc.*'. Module '<root>' cannot depend on 'pennylane.ftqc'.
```

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**