Fixing jitting with device prng key (#6899)
**Context:**

When setting a prng key on the device and using jax-jit, we get a leaked
tracer error.

**Description of the Change:**

If the user specifies a `PRNGKey` as a seed, we assume they want to jit
end-to-end.

We raise a warning during `DefaultQubit.execute` if we are doing native
jitting and have a decently sized batch. This avoids the huge
compilation overhead problem.

**Benefits:**

No more leaked tracers. Users can jit end to end by specifying a
PRNGKey.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>