Raise `RuntimeWarning` if jax > 0.4.28 is installed (#6864)
Jax is not compatible with pennylane for versions > 0.4.28 (confirmed by
the test logs of [this
PR](https://github.com/PennyLaneAI/pennylane/pull/6768)). This PR raises
a `RuntimeWarning` if version > 0.4.28 is installed:
* In `QNode.__init__` if the specified interface resolves to `JAX` or
`JAX_JIT`
* In `execute` if the original interface was `"auto"` or `None` and the
resolved interface is `JAX` or `JAX_JIT`. I don't raise a warning if the
original interface was jax/jax-jit because if that is the case,
`QNode.__init__` should've already raised a warning.
* This is done by adding a `_validate_jax_version` private kwarg to
`qml.math.get_canonical_interface_name`, which is called by
`QNode.__init__`, and `qml.workflow.resolution._resolve_interface` (used
by `qml.execute`), so that both those entry points can raise warnings if
appropriate.

Possible drawbacks:
None. The only annoying is that our CI uses supported versions of jax,
so we can't get coverage for the runtime warning. I've added a `#
pragma: no cover` to work around this, but I'm not particularly happy
about it.

[sc-82570]